requirements.txt
python-dotenv
transformers
fla-core
torch
